Service security audit

+++ contracts/TalentLayerService.sol
@@ -70,7 +70,7 @@ contract TalentLayerService is Initializable, ERC2771RecipientUpgradeable, UUPSU
         uint256 ownerId;
         address rateToken;
         uint256 rateAmount;
-        uint16 platformId;
+        uint256 platformId;
         string dataUri;
         uint256 expirationDate;
     }
@@ -127,7 +127,7 @@ contract TalentLayerService is Initializable, ERC2771RecipientUpgradeable, UUPSU
         ProposalStatus status,
         address rateToken,
         uint256 rateAmount,
-        uint16 platformId,
+        uint256 platformId,
         uint256 expirationDate
     );
 
@@ -343,7 +343,7 @@ contract TalentLayerService is Initializable, ERC2771RecipientUpgradeable, UUPSU
         uint256 _serviceId,
         address _rateToken,
         uint256 _rateAmount,
-        uint16 _platformId,
+        uint256 _platformId,
         string calldata _dataUri,
         uint256 _expirationDate,
         bytes calldata _signature
@@ -584,7 +584,7 @@ contract TalentLayerService is Initializable, ERC2771RecipientUpgradeable, UUPSU
         uint256 _serviceId,
         address _rateToken,
         uint256 _rateAmount,
-        uint16 _platformId,
+        uint256 _platformId,
         string calldata _dataUri,
         bytes calldata _signature
     ) private view {
